requirements.txt
datasets
Flask
Flask-Cors
Gunicorn
python-okx
autobahn
tqdm
requests
pandas
firebase-admin
newsapi-python
gnews
scikit-learn
transformers
seaborn
wordcloud
fastparquet
reactivex
apscheduler
flask_socketio
websocket-client
